+++ geofileops/fileops.py
@@ -95,85 +95,6 @@ PRJ_EPSG_31370 = (
 )
 
 
-def concat(
-    input_paths: Iterable[Union[str, "os.PathLike[Any]"]],
-    output_path: Union[str, "os.PathLike[Any]"],
-    input_layers: Iterable[str | None] | None = None,
-    output_layer: str | None = None,
-    columns: Iterable[str] | None = None,
-    explodecollections: bool = False,
-    spatial_index: bool | None = None,
-    force: bool = False,
-):
-    """Concatenate multiple geofiles into one output geofile.
-
-    Args:
-        input_paths (Iterable[PathLike]): the paths to the files to concatenate.
-        output_path (PathLike): the path to the output file.
-        input_layers (Iterable[str | None], optional): the layer names to use in the
-            input files. The layer names don't need to be specified for input files that
-            only contain a single layer. Defaults to None.
-        output_layer (str, optional): the layer name to use in the output file. If not
-            specified, the default layer name is used. Defaults to None.
-        columns (Iterable[str], optional): the columns to keep in the output file.
-            If not specified, all columns are kept. Defaults to None.
-        explodecollections (bool, optional): True to explode geometry collections
-            into separate features. Defaults to False.
-        spatial_index (bool, optional): True to create a spatial index on the
-            destination file/layer. If None, the default behaviour by gdal for that file
-            type is respected. Defaults to None.
-        force (bool, optional): True to overwrite the output file if it already exists.
-    """
-    # Validate + cleanup input parameters
-    input_paths = list(input_paths)
-    if input_layers is None:
-        input_layers = [None] * len(input_paths)
-    else:
-        input_layers = list(input_layers)
-        if len(input_layers) != len(input_paths):
-            raise ValueError(
-                "src_layers must have the same length as the src file list if specified"
-            )
-    if _vsi_exists(output_path) and not force:
-        logger.info(f"Destination file already exists, so stop: {output_path}")
-        return
-
-    # Concat all files
-    tmp_dir = _io_util.create_tempdir("concat")
-    tmp_dst = tmp_dir / Path(output_path).name
-    is_first = True
-    for src_path, src_layer in zip(input_paths, input_layers):
-        if is_first:
-            force_local = force
-            write_mode = "create"
-        else:
-            force_local = False
-            write_mode = "append"
-
-        copy_layer(
-            src=src_path,
-            dst=tmp_dst,
-            write_mode=write_mode,
-            src_layer=src_layer,
-            dst_layer=output_layer,
-            columns=columns,
-            explodecollections=explodecollections,
-            create_spatial_index=False,
-            force=force_local,
-        )
-
-        if is_first:
-            is_first = False
-
-    # Add a spatial index if needed
-    if spatial_index is None:
-        spatial_index = _geofileinfo.get_geofileinfo(tmp_dst).default_spatial_index
-    if spatial_index:
-        create_spatial_index(tmp_dst, output_layer)
-
-    move(tmp_dst, output_path)
-
-
 def listlayers(
     path: Union[str, "os.PathLike[Any]"], only_spatial_layers: bool = True
 ) -> list[str]:
@@ -2789,6 +2710,11 @@ def copy_layer(
                 layer already exists. If not, the file and/or layer will be created.
                 If the file already contains layers named differently than the default
                 layer name for the file, `dst_layer` becomes mandatory.
+            - "append_add_fields": append the source layer to the destination layer,
+                adding fields from the source layer that are not present in the
+                destination layer. If the layer doesn't exist, it will be created.
+                If the file already contains layers named differently than the default
+                layer name for the file, `dst_layer` becomes mandatory.
 
         src_crs (Union[str, int], optional): an epsg int or anything supported
             by the OGRSpatialReference.SetFromUserInput() call, which includes
@@ -2869,7 +2795,8 @@ def copy_layer(
         )
         write_mode = "append"
 
-    # Determine the access mode
+    # Determine the access mode and whether to add missing fields when appending
+    add_fields = False
     access_mode = _determine_access_mode(dst, dst_layer, write_mode, force)
     if access_mode is None:
         # The file/layer exists already and force is false, so we can return
@@ -2881,6 +2808,8 @@ def copy_layer(
         # As we will actually be appending to an existing layer, the layer
         # creation option regarding spatial index creation should not be passed.
         create_spatial_index = None
+        if write_mode == "append_add_fields":
+            add_fields = True
 
     # Check/clean input params
     if isinstance(columns, str):
@@ -2906,6 +2835,7 @@ def copy_layer(
     # copy, it is faster to copy the data directly in sqlite instead of via GDAL.
     if (
         access_mode == "append"
+        and not add_fields
         and Path(src).suffix.lower() == ".gpkg"
         and Path(dst).suffix.lower() == ".gpkg"
         and not sql_stmt
@@ -2922,7 +2852,6 @@ def copy_layer(
         and ConfigOptions.copy_layer_sqlite_direct
     ):
         # TODO: sql_stmt?, access_mode="create", create_spatial_index, dst_crs?
-        # TODO: create gfo config option to disable?
         try:
             name = src_layername if src_layername is not None else get_only_layer(src)
             preserve_fid_local = preserve_fid if preserve_fid is not None else False
@@ -3011,6 +2940,7 @@ def copy_layer(
         options=options,
         preserve_fid=preserve_fid,
         dst_dimensions=dst_dimensions,
+        add_fields=add_fields,
     )
     _ogr_util.vector_translate_by_info(info=translate_info)
 
@@ -3105,7 +3035,7 @@ def _determine_access_mode(
         else:
             return "update"
 
-    elif write_mode == "append":
+    elif write_mode in ("append", "append_add_fields"):
         layers = try_listlayers(dst, only_spatial_layers=False)
         if layers is None:
             # The file doesn't seem to exist yet... just continue, the file and layer
@@ -3119,16 +3049,17 @@ def _determine_access_mode(
             dst_layer = get_default_layer(dst)
             if dst_layer not in layers:
                 raise ValueError(
-                    "dst_layer is required when write_mode is 'append' and "
-                    "there are already other layers than the default layername."
+                    "dst_layer is required when write_mode is 'append' or "
+                    "'append_add_fields' and there are already other layers than the "
+                    "default layername."
                 )
             return "append"
 
         elif dst_layer is None:
             # No dst_layer specified and multiple layers: raise error
             raise ValueError(
-                "dst_layer is required when write_mode is 'append' and "
-                "there are multiple other layers."
+                "dst_layer is required when write_mode is 'append' or "
+                "'append_add_fields' and there are multiple other layers."
             )
         elif dst_layer in layers:
             return "append"