+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 
 ### Improvements
 
+- Add `concat` function (#746, #747)
+- Add `write_mode="append_add_fields"` option to `copy_layer` (#750)
 - Improve performance of clip with a complex clip layer (#740)
 - Improve performance of most operations by using a direct gpkg to gpkg append via
   sqlite where possible (#728)